+++ jest.config.js
@@ -32,6 +32,10 @@ module.exports = {
     "src/app/core/spi-impl/hybrid-key-storage.provider.ts",
     "src/app/features/hybrid-onboarding/hybrid-onboarding.page.ts",
     "src/app/features/signing/sign-prompt.component.ts",
+    "src/app/features/hybrid-keymanager/hybrid-kdf.const.ts",
+    "src/app/features/hybrid-keymanager/unwrap.service.ts",
+    "src/app/features/hybrid-keymanager/sign.service.ts",
+    "src/app/features/hybrid-keymanager/sign.api.ts",
   ],
   coveragePathIgnorePatterns: [
     '<rootDir>/node_modules/',